feat jwt관련 클래스 및 SecurityConfig 클래스 구현+ 회원가입 및 로그인 로직 구현

+++ fpsgame/src/main/java/Capstone/fpsgame/domain/Weapon.java
@@ -15,10 +15,10 @@ public class Weapon {
     private Long id;
     private String name;
     @Column(name="attack_power")
-    private Long attackPower;
+    private float attackPower;
     @Column(name="ammo_capacity")
     private int ammoCapacity;
-    private Long weight;
+    private float weight;
     @OneToOne(mappedBy = "weapon")
     private UserInfo userInfo;
